refactor: integrated global kanban view everywhere
