+++ apps/app/components/core/index.ts
@@ -1,5 +1,5 @@
-export * from "./not-authorized-view";
 export * from "./bulk-delete-issues-modal";
 export * from "./existing-issues-list-modal";
 export * from "./image-upload-modal";
-export * from "./view";
+export * from "./issues-filter-view";
+export * from "./not-authorized-view";
