+++ apps/app/pages/[workspaceSlug]/projects/[projectId]/cycles/[cycleId].tsx
@@ -12,10 +12,10 @@ import AppLayout from "layouts/app-layout";
 import { IssueViewContextProvider } from "contexts/issue-view.context";
 // components
 import CyclesListView from "components/project/cycles/list-view";
-import CyclesBoardView from "components/project/cycles/board-view";
 import { CreateUpdateIssueModal, DeleteIssueModal } from "components/issues";
 import { ExistingIssuesListModal, IssuesFilterView } from "components/core";
 import CycleDetailSidebar from "components/project/cycles/cycle-detail-sidebar";
+import { AllBoards } from "components/core/board-view/all-boards";
 // services
 import issuesServices from "services/issues.service";
 import cycleServices from "services/cycles.service";
@@ -242,12 +242,11 @@ const SingleCycle: React.FC<UserAuth> = (props) => {
                 setPreloadedData={setPreloadedData}
                 userAuth={props}
               />
-              <CyclesBoardView
+              <AllBoards
+                type="cycle"
                 issues={cycleIssuesArray ?? []}
-                openCreateIssueModal={openCreateIssueModal}
-                openIssuesListModal={openIssuesListModal}
                 handleDeleteIssue={setDeleteIssue}
-                setPreloadedData={setPreloadedData}
+                openIssuesListModal={openIssuesListModal}
                 userAuth={props}
               />
             </div>
